+++ src/test/java/racingcar/RacingTest.java
@@ -2,6 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Racing;
+import racingcar.domain.RacingResult;
 
 import java.util.Arrays;
 import java.util.List;
